Move landscape preprocessing to a new pass
