+++ projects/PathosEngine/PathosEngine.vcxproj.filters
@@ -423,6 +423,9 @@
     <ClCompile Include="src\pathos\scene\landscape_component.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="src\pathos\render\landscape_rendering.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\pathos\mesh\geometry.h">
@@ -927,6 +930,9 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="src\pathos\scene\landscape_component.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="src\pathos\render\landscape_rendering.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
