+++ projects/PathosEngine/src/pathos/render/gbuffer_pass.h
@@ -28,7 +28,6 @@ namespace pathos {
 	private:
 		GLuint fbo = 0;
 		UniformBuffer uboPerObject;
-		UniformBuffer uboLandscapeCulling;
 
 		MeshGeometry* fullscreenQuad = nullptr;
